Player Pal customisation front sprite
Moves kyphii's dynpal to custom sprites. Missing OWs and backsprites. Skin, Hair and Clothes all seem to map to pal slots 5-9, need to investigate.

+++ include/global.h
@@ -562,9 +562,9 @@ struct SaveBlock2
 #endif //FREE_RECORD_MIXING_HALL_RECORDS
     /*0x624*/ u16 contestLinkResults[CONTEST_CATEGORIES_COUNT][CONTESTANT_COUNT];
     /*0x64C*/ struct BattleFrontier frontier;
-    u8 dynPalPartAPreset;
-    u8 dynPalPartBPreset;
-    u8 dynPalPartCPreset;
+    u16 dynPalPartAPreset;
+    u16 dynPalPartBPreset;
+    u16 dynPalPartCPreset;
 }; // sizeof=0xF2C
 
 extern struct SaveBlock2 *gSaveBlock2Ptr;
